Version 2.5.0D
파일처리 변수 int -> qint64 변경으로 오버 플로어 방지

+++ BerEditor/enc_dec_thread.cpp
@@ -70,7 +70,7 @@ void EncDecThread::run()
     int ret = 0;
     int nRead = 0;
     int nPartSize = berApplet->settingsMgr()->fileReadSize();
-    int nReadSize = 0;
+    qint64 nReadSize = 0;
     int nLeft = 0;
     int nOffset = 0;
     BIN binPart = {0,0};
