+++ BerEditor/hash_thread.cpp
@@ -33,7 +33,7 @@ void HashThread::run()
     int ret = 0;
     int nRead = 0;
     int nPartSize = berApplet->settingsMgr()->fileReadSize();
-    int nReadSize = 0;
+    qint64 nReadSize = 0;
     int nLeft = 0;
     int nOffset = 0;
     int nUpdateCnt = 0;
