+++ BerEditor/sign_verify_dlg.h
@@ -58,7 +58,7 @@ private slots:
     void fileRunThread();
     void startTask();
     void onTaskFinished();
-    void onTaskUpdate( int nUpdate );
+    void onTaskUpdate( qint64 nUpdate );
 
     void checkCertGroup();
