+++ BerEditor/sign_verify_thread.cpp
@@ -41,7 +41,7 @@ void SignVerifyThread::run()
     int ret = 0;
     int nRead = 0;
     int nPartSize = berApplet->settingsMgr()->fileReadSize();
-    int nReadSize = 0;
+    qint64 nReadSize = 0;
     int nLeft = 0;
     int nOffset = 0;
